python-ai/requirements.txt
# BioVerse Python AI Service Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI/ML Stack
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.0
transformers==4.35.0
langchain==0.1.0
langchain-community==0.0.10

# Ollama Integration
ollama==0.1.7
httpx==0.25.2

# Healthcare & Medical
fhir.resources==7.0.2
pydicom==2.4.3
biopython==1.81

# Data Processing & Analysis
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# 3D Visualization & Digital Twins
vtk==9.3.0
pyvista==0.43.0
open3d==0.18.0
trimesh==4.0.5

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
imageio==2.33.1

# Real-time & WebSocket
websockets==12.0
python-socketio==5.10.0

# Async & Concurrency
asyncio==3.4.3
aioredis==2.0.1
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment
python-dotenv==1.0.0